backend/requirements.txt
psycopg[binary]
sqlalchemy
pgvector
streamlit
fastapi
python-dotenv